# src/swaggerOutput.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "BlockInv",
    "description": "BlockInv REST API Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/",
      "description": "Local Host"
    }
  ],
  "paths": {
...
    "/earning/all": {
      "get": {
        "tags": [
          "User Earnings"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EarningsResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/EarningsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/earning/{id}": {
      "get": {
        "tags": [
          "User Earnings"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EarningResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/EarningResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/admin/earnings/all": {
      "get": {
        "tags": [
          "Admin - Earnings"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EarningssResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/EarningssResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/earnings/user/{uid}": {
      "get": {
        "tags": [
          "Admin - Earnings"
        ],
        "description": "",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EarningssResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/EarningssResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/earnings/{uid}": {
      "post": {
        "tags": [
          "Admin - Earnings"
        ],
        "description": "",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Earnings added to the user balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EarningRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/EarningRequest"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EarningRequest": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "string",
            "example": "number"
          },
          "plan_id": {
            "type": "string",
            "example": "string"
          },
          "investment_id": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "amount",
          "plan_id",
          "investment_id"
        ],
        "xml": {
          "name": "EarningRequest"
        }
      },
...
      "Response": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "string"
          }
        },
        "xml": {
          "name": "Response"
        }
      },
...
      "EarningsResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "string"
                },
                "uid": {
                  "type": "string",
                  "example": "string"
                },
                "amount": {
                  "type": "string",
                  "example": "number"
                },
                "plan_name": {
                  "type": "string",
                  "example": "string"
                },
                "plan_id": {
                  "type": "string",
                  "example": "string"
                },
                "investment_id": {
                  "type": "string",
                  "example": "string"
                },
                "createdAt": {
                  "type": "string",
                  "example": "number"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "number"
                }
              }
            }
          }
        },
        "xml": {
          "name": "EarningsResponse"
        }
      },
      "EarningResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "string"
              },
              "uid": {
                "type": "string",
                "example": "string"
              },
              "amount": {
                "type": "string",
                "example": "number"
              },
              "plan_name": {
                "type": "string",
                "example": "string"
              },
              "plan_id": {
                "type": "string",
                "example": "string"
              },
              "investment_id": {
                "type": "string",
                "example": "string"
              },
              "createdAt": {
                "type": "string",
                "example": "number"
              },
              "updatedAt": {
                "type": "string",
                "example": "number"
              }
            }
          }
        },
        "xml": {
          "name": "EarningResponse"
        }
      },
...
      "InvalidToken": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid Token."
          }
        },
        "xml": {
          "name": "InvalidToken"
        }
      },
...
      "NotExists": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Does not exists."
          }
        },
        "xml": {
          "name": "NotExists"
        }
      },
      "InvalidID": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid ID."
          }
        },
        "xml": {
          "name": "InvalidID"
        }
      },
...
    }
  },
...
}